update: refactor user service to use common library package
