+++ user-service/pom.xml
@@ -93,11 +93,11 @@
 			<artifactId>jaxb-api</artifactId>
 			<version>2.3.1</version>
 		</dependency>
-		<!-- Shared entities library for Fizzy Budget -->
+		<!-- Shared common library for Fizzy Budget -->
 		<dependency>
 			<groupId>com.huiun.fizzybudget</groupId>
-			<artifactId>shared-entities</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
+			<artifactId>common</artifactId>
+			<version>0.0.2</version>
 		</dependency>
 	</dependencies>
