+++ user-service/src/main/java/com/huiun/fizzybudget/userservice/UserServiceApplication.java
@@ -2,10 +2,8 @@ package com.huiun.fizzybudget.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 @SpringBootApplication
-@EntityScan(basePackages = "com.huiun.fizzybudget.sharedentities") // Adjust the package to your shared entities package
 public class UserServiceApplication {
 
 	public static void main(String[] args) {
