+++ user-service/src/main/java/com/huiun/fizzybudget/userservice/config/SecurityConfig.java
@@ -1,6 +1,7 @@
 package com.huiun.fizzybudget.userservice.config;
 
-import com.huiun.fizzybudget.userservice.security.JWTAuthenticationFilter;
+import com.huiun.fizzybudget.common.security.JWTAuthenticationFilter;
+import com.huiun.fizzybudget.common.security.JWTTokenProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
