--- user-service/src/main/java/com/huiun/fizzybudget/userservice/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package com.huiun.fizzybudget.userservice.repository;
-
-import com.huiun.fizzybudget.sharedentities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUserId(Long userId);
-
-    Optional<User> findByUsername(String username);
-
-    Optional<User> findByEmail(String email);
-}
