--- user-service/src/main/java/com/huiun/fizzybudget/userservice/service/CustomUserDetailsService.java
@@ -1,24 +1,0 @@
-package com.huiun.fizzybudget.userservice.service;
-
-import com.huiun.fizzybudget.sharedentities.User;
-import com.huiun.fizzybudget.userservice.repository.UserRepository;
-import com.huiun.fizzybudget.userservice.security.CustomUserDetails;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-         User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
-        return new CustomUserDetails(user);
-    }
-}
